+++ lib/http.dart
@@ -1,0 +1,9 @@
+/// HTTP 1.1 / HTTP 2
+///
+/// HTTP multi-version client and server classes.
+library http;
+
+export 'src/http/http_client.dart';
+export 'src/http/http_server.dart';
+export 'src/http/http_request.dart';
+export 'src/http/http_response.dart';