+++ lib/src/http/http_response.dart
@@ -1,0 +1,8 @@
+class HttpResponse {
+  final Uri requestUrl;
+  final int statusCode;
+  final Map<String, List<String>> headers;
+  final Stream<List<int>> bodyData;
+
+  HttpResponse(this.requestUrl, this.statusCode, this.headers, this.bodyData);
+}